fix(acis): native-genus ASM synthesis - the authored-form census, byte-verified
Third BricsCAD verdict on the ASM port (2026-09-24, cylinder_hostchain
probe): the fix/sat-validate-before-sab forms were themselves never
strict-loader-verified. Byte-decoding the authored specimens (the six
parametric 2018 fixtures, uniform) against the ported synthesis showed
every synthesized form diverging from the native genus:

  asmheader:  native = attribute $-1 + INTEGER(-1) + STRING("232.6.0.65535")
              port  = attribute $-1 + STRING("208.0.4.7009") - missing the
              INTEGER(-1) token, wrong version string
  transform:  native = attribute $-1 + INTEGER(-1) + four DIRECTION(0x14)
              rows (3x4 placement) + DOUBLE(1.0) determinant + TRUE x3
              port  = attribute $-1 + ASM_LONG_STRING(0x12) text blob -
              an unknown-tag record that desyncs the stream at the
              transform, the "Invalid input" signature
  header:     native = [num_records 0, num_bodies 2, history/flags 4]
              (and 26/24/12 at versions 21200/21500/21800 - the word is
              version-dependent, not a bool)
              port  = [real count, 1, 0/1]
  order:      native leads asmheader, body, lump, transform, shell;
              the port's blanket reversal put body-LAST for new_body-
              built documents (body-first assemblies) and the transform
              at the tail

The reader already parsed all of this correctly (corpus 0/0), so the
writer was made to emit what the reader knows:

- SabReader/SabWriter/round-trip: SatHeader.has_history becomes the
  raw u32 word (retained verbatim; text-parsed docs carry 0/1). The
  native Box_2018 SAB now round-trips SabReader -> write_asm
  BYTE-IDENTICALLY (asm_roundtrip_probe; the previous divergence was
  exactly the 4->1 bool truncation).
- to_asm_structure, rewritten to the census: native asmheader and
  transform forms; layout asmheader, body, lump, transform, shell,
  then the remaining records in their existing order (both assembly
  genera - cadkernel body-last, new_body body-first - restore
  body-first); the body's LAST pointer token (the transform slot in
  both the 5-token native and 4-token new_body shapes) links to the
  transform - the document's OWN transform when it carries one
  (placed at the canonical slot; the regression gate caught the
  duplicate-transform bug), the synthetic identity otherwise; header
  stamped [22300, 0, 2, 4] with product strings "Autodesk AutoCAD" /
  "ASM 232.6.0.65535 NT" and spatial_resolution 1.0.
- modern_face_and_transform_boolean_roles_roundtrip: the text
  round-trip asserts the raw word verbatim ("21200 2 1 26" - no bool
  truncation).

Gates: cargo test / cargo test --features serde green (1315 lib +
integration + doc); gen_all 30 OK, canonical md5 moves to
3493190b1c988642f7cdd4887508848f (the ASM wrapper's native forms);
gold-harness corpus 280 files 0/0; native round-trip byte-identical.

+++ src/entities/acis/writer.rs
@@ -26,7 +26,7 @@ impl SatWriter {
             doc.header.version.sat_version_number(),
             num_records_out,
             doc.header.num_bodies,
-            u32::from(doc.header.has_history)
+            doc.header.has_history
         ));
 
         // Header line 2: product info
